AutoTimerAddon.py: add prototype function to advertise addons
**************
To not make things even more overly complicated, the name should NOT be
localized (actually this is true for any plugin :P), but we now REQUIRE
the name to be static for proper recognition by external software.
Due to a lack of versioning in this single parameter the name can be
e.g. MyAwesomeAddon5 or MyAwesomeAddon6 to advertise the addon version.
**************