[EPGSearch] This is a first version of using AT-filter for EPGSearch. -EPGSearchAT-class was derived from AutoTimer-class and uses his own filename. -The search results are displayed in the regular EPGSearch-Screen
To do: there is no checking if AutoTimer is installed or not. The default settings of this filter search are the regular AutoTimer-settings. If we would change this we would have a lot of source code. There is no option to disable the filter-search yet.

I'm still not sure what the advantage of this search method is. Does this really makes sense to use filter? :)
Anyway, Sven can play with it in this branch. ;)